düzenlendi.

+++ OtelRezervasyonSistemi/OtelRezervasyonSistemi/FrmMusteriler.cs
@@ -113,78 +113,78 @@ namespace OtelRezervasyonSistemi
             baglanti.Close();
             verilergöster();
 
-            /*if (TxtOdaNo.Text=="101")
+            if (TxtOdaNo.Text=="101")
             {
                 baglanti.Open();
-                SqlCommand komut = new SqlCommand("delete from Oda101", baglanti);
-                komut.ExecuteNonQuery();
+                SqlCommand komut1 = new SqlCommand("delete from Oda101", baglanti);
+                komut1.ExecuteNonQuery();
                 baglanti.Close();
                 verilergöster();
             }
             if (TxtOdaNo.Text == "102")
             {
                 baglanti.Open();
-                SqlCommand komut = new SqlCommand("delete from Oda102", baglanti);
-                komut.ExecuteNonQuery();
+                SqlCommand komut1 = new SqlCommand("delete from Oda102", baglanti);
+                komut1.ExecuteNonQuery();
                 baglanti.Close();
                 verilergöster();
             }
             if (TxtOdaNo.Text == "103")
             {
                 baglanti.Open();
-                SqlCommand komut = new SqlCommand("delete from Oda103", baglanti);
-                komut.ExecuteNonQuery();
+                SqlCommand komut1 = new SqlCommand("delete from Oda103", baglanti);
+                komut1.ExecuteNonQuery();
                 baglanti.Close();
                 verilergöster();
             }
             if (TxtOdaNo.Text == "104")
             {
                 baglanti.Open();
-                SqlCommand komut = new SqlCommand("delete from Oda104", baglanti);
-                komut.ExecuteNonQuery();
+                SqlCommand komut1 = new SqlCommand("delete from Oda104", baglanti);
+                komut1.ExecuteNonQuery();
                 baglanti.Close();
                 verilergöster();
             }
             if (TxtOdaNo.Text == "105")
             {
                 baglanti.Open();
-                SqlCommand komut = new SqlCommand("delete from Oda105", baglanti);
-                komut.ExecuteNonQuery();
+                SqlCommand komut1 = new SqlCommand("delete from Oda105", baglanti);
+                komut1.ExecuteNonQuery();
                 baglanti.Close();
                 verilergöster();
             }
             if (TxtOdaNo.Text == "106")
             {
                 baglanti.Open();
-                SqlCommand komut = new SqlCommand("delete from Oda106", baglanti);
-                komut.ExecuteNonQuery();
+                SqlCommand komut1 = new SqlCommand("delete from Oda106", baglanti);
+                komut1.ExecuteNonQuery();
                 baglanti.Close();
                 verilergöster();
             }
             if (TxtOdaNo.Text == "107")
             {
                 baglanti.Open();
-                SqlCommand komut = new SqlCommand("delete from Oda107", baglanti);
-                komut.ExecuteNonQuery();
+                SqlCommand komut1 = new SqlCommand("delete from Oda107", baglanti);
+                komut1.ExecuteNonQuery();
                 baglanti.Close();
                 verilergöster();
             }
             if (TxtOdaNo.Text == "108")
             {
                 baglanti.Open();
-                SqlCommand komut = new SqlCommand("delete from Oda108", baglanti);
-                komut.ExecuteNonQuery();
+                SqlCommand komut1 = new SqlCommand("delete from Oda108", baglanti);
+                komut1.ExecuteNonQuery();
                 baglanti.Close();
                 verilergöster();
             }
             if (TxtOdaNo.Text == "109")
             {
                 baglanti.Open();
-                SqlCommand komut = new SqlCommand("delete from Oda109", baglanti);
-                komut.ExecuteNonQuery();
+                SqlCommand komut1 = new SqlCommand("delete from Oda109", baglanti);
+                komut1.ExecuteNonQuery();
                 baglanti.Close();
                 verilergöster();
-            }*/
+            }
 
         }
